new whirlpool data structure
