+++ src/admin/public/types.ts
@@ -1,28 +1,15 @@
+import { TickSpacing } from "@orca-so/whirlpool-client-sdk";
 import { Address, BN, Provider } from "@project-serum/anchor";
 
 export type InitPoolTxParam = {
   initSqrtPrice: BN;
   tokenMintA: Address;
   tokenMintB: Address;
-  stable: boolean;
-};
-
-export type CollectProtocolFeesTxParam = {
-  poolAddress: Address;
+  tickSpacing: TickSpacing;
 };
 
 export type SetFeeAuthorityTxParam = {
   newFeeAuthority: Address;
-};
-
-export type SetFeeRateTxParam = {
-  poolAddress: Address;
-  feeRate: number;
-};
-
-export type SetProtocolFeeRateTxParam = {
-  poolAddress: Address;
-  protocolFeeRate: number;
 };
 
 export type SetCollectProtocolFeesAuthorityTxParam = {
@@ -31,19 +18,16 @@ export type SetCollectProtocolFeesAuthorityTxParam = {
 
 export type InitRewardTxParam = {
   rewardAuthority: Address;
-  poolAddress: Address;
   rewardMint: Address;
   rewardIndex: number;
 };
 
 export type SetRewardAuthorityTxParam = {
-  poolAddress: Address;
   newRewardAuthority: Address;
   rewardIndex: number;
 };
 
 export type SetRewardEmissionsTxParam = {
-  poolAddress: Address;
   rewardIndex: number;
   emissionsPerSecondX64: BN;
 };
